Deduplicated ActionRow, batch TransactionAction, and ActionRow with minimal details with ActionRowBlock component
