+++ frontend/src/components/transactions/ActionRow.tsx
@@ -1,24 +1,16 @@
 import React from "react";
-import { Row, Col } from "react-bootstrap";
 
-import AccountLink from "../utils/AccountLink";
-import ExecutionStatus from "../utils/ExecutionStatus";
-import TransactionLink from "../utils/TransactionLink";
-import Timer from "../utils/Timer";
 import actionIcons from "./ActionIcons";
 import ActionMessage from "./ActionMessage";
+import ActionRowBlock, { ViewMode, DetalizationMode } from "./ActionRowBlock";
 import * as T from "../../libraries/explorer-wamp/transactions";
 
-export type ViewMode = "sparse" | "compact";
-export type DetalizationMode = "detailed" | "minimal";
 export interface Props {
   action: T.Action | keyof T.Action;
   transaction: T.Transaction;
   viewMode: ViewMode;
   detalizationMode: DetalizationMode;
   className: string;
-  batch?: boolean;
-  detail?: boolean;
 }
 
 export interface State {}
@@ -50,74 +42,21 @@ export default class extends React.Component<Props, State> {
     }
 
     const ActionIcon = actionIcons[actionKind];
-    let actionRow;
-    if (detalizationMode === "minimal") {
-      actionRow = (
-        <Row noGutters className={`action-${viewMode}-row mx-0 ${className}`}>
-          <Col xs="auto" className="actions-icon-col">
-            <div className="action-row-img">{ActionIcon && <ActionIcon />}</div>
-          </Col>
-          <Col className="action-row-details">
-            <Row noGutters>
-              <Col md="8" xs="7">
-                <Row noGutters>
-                  <Col className="action-row-title">
-                    <ActionMessage
-                      transaction={transaction}
-                      actionKind={actionKind}
-                      actionArgs={actionArgs}
-                    />
-                  </Col>
-                </Row>
-              </Col>
-            </Row>
-          </Col>
-        </Row>
-      );
-    } else {
-      actionRow = (
-        <Row noGutters className={`action-${viewMode}-row mx-0 ${className}`}>
-          <Col xs="auto">
-            <div className="action-row-img">{ActionIcon && <ActionIcon />}</div>
-          </Col>
-          <Col className="action-row-details">
-            <Row noGutters>
-              <Col md="8" xs="7">
-                <Row noGutters>
-                  <Col className="action-row-title">
-                    <ActionMessage
-                      transaction={transaction}
-                      actionKind={actionKind}
-                      actionArgs={actionArgs}
-                    />
-                  </Col>
-                </Row>
-                <Row noGutters>
-                  <Col className="action-row-text">
-                    by <AccountLink accountId={transaction.signerId} />
-                  </Col>
-                </Row>
-              </Col>
-              <Col md="4" xs="5" className="ml-auto text-right">
-                <Row>
-                  <Col className="action-row-txid">
-                    <TransactionLink transactionHash={transaction.hash} />
-                  </Col>
-                </Row>
-                <Row>
-                  <Col className="action-row-timer">
-                    <span className="action-row-timer-status">
-                      <ExecutionStatus status={transaction.status} />
-                    </span>{" "}
-                    <Timer time={transaction.blockTimestamp} />
-                  </Col>
-                </Row>
-              </Col>
-            </Row>
-          </Col>
-        </Row>
-      );
-    }
-    return <>{actionRow}</>;
+    return (
+      <ActionRowBlock
+        viewMode={viewMode}
+        detalizationMode={detalizationMode}
+        className={className}
+        transaction={transaction}
+        icon={ActionIcon && <ActionIcon />}
+        title={
+          <ActionMessage
+            transaction={transaction}
+            actionKind={actionKind}
+            actionArgs={actionArgs}
+          />
+        }
+      />
+    );
   }
 }
